yay for macro generated Queryable impls

+++ gdcf_diesel/src/profile.rs
@@ -25,27 +25,27 @@ diesel_stuff! {
         (creator_points, creator_points, u16),
         (index_10, index_10, String),
         (index_11, index_11, String),
-        (secret_coints, secret_coins, u16),
+        (secret_coints, secret_coins, u8),
         (account_id, account_id, u64),
         (user_coins, user_coins, u16),
         (index_18, index_18, String),
         (index_19, index_19, String),
         (youtube_url, youtube_url, Option<String>),
         (cube_index, cube_index, u16),
-        (ship_index, ship_index, u16),
-        (ball_index, ball_index, u16),
-        (ufo_index, ufo_index, u16),
-        (wave_index, wave_index, u16),
-        (robot_index, robot_index, u16),
+        (ship_index, ship_index, u8),
+        (ball_index, ball_index, u8),
+        (ufo_index, ufo_index, u8),
+        (wave_index, wave_index, u8),
+        (robot_index, robot_index, u8),
         (has_glow, has_glow, bool),
         (index_29, index_29, String),
         (global_rank, global_rank, Option<u32>),
         (index_31, index_31, String),
-        (spider_index, spider_index, u16),
+        (spider_index, spider_index, u8),
         (twitter_url, twitter_url, Option<String>),
         (twitch_url, twitch_url, Option<String>),
         (diamonds, diamonds, u16),
-        (death_effect_index, death_effect_index, u16),
+        (death_effect_index, death_effect_index, u8),
         (index_49, index_49, String),
         (index_50, index_50, String)
     }
@@ -54,7 +54,7 @@ meta_table!(profile_meta, user_id);
 
 store_simply!(User, profile, profile_meta, user_id);
 lookup_simply!(User, profile, profile_meta, user_id);
-
+/*
 impl<DB: Backend> Queryable<SqlType, DB> for Wrapped<User>
 where
     Row: FromSqlRow<SqlType, DB>,
@@ -65,3 +65,4 @@ where
         unimplemented!()
     }
 }
+*/
